+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Version History
 
+## 1.1.0 Jun 19, 2024
+
+* Added multilabel classification.
+* Added regression.
+* Change from aggregating probabilities to aggregating logits/scores.
+
 ## 1.0.1 Feb 26, 2024
 
 * Add gradient accumulation by @jstremme in https://github.com/mim-solutions/bert_for_longer_texts/pull/23.